requirements.txt
python-dotenv==1.2.2
requests==2.32.5
pandas==3.0.1
